requirements.txt
annotated-types==0.6.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
